@@ -1,18 +1,18 @@
 source 'https://rubygems.org'
-ruby '2.6.10'
+ruby '3.2.1'
 
 # Bundle edge Rails instead: gem 'rails', github: 'rails/rails'
-gem 'rails', '4.2.8'
+gem 'rails', '7.0.1'
 # Use sqlite3 as the database for Active Record
 gem 'sqlite3', :group => [:development, :test]
 # Use SCSS for stylesheets
-gem 'sass-rails', '4.0.2'
+gem 'sass-rails', '6.0.0'
 # Use Uglifier as compressor for JavaScript assets
 gem 'uglifier', '>= 1.3.0'
 # See https://github.com/rails/execjs#readme for more supported runtimes
 # gem 'therubyracer', platforms: :ruby
 gem 'bootstrap-sass',       '3.2.0.4'
-gem 'activerecord', '4.2.8'
+gem 'activerecord', '7.0.1'
 # Use jquery as the JavaScript library
 gem 'jquery-rails'
 gem 'jquery-turbolinks'
@@ -20,21 +20,20 @@ gem 'jquery-turbolinks'
 gem 'turbolinks'
 # Build JSON APIs with ease. Read more: https://github.com/rails/jbuilder
 gem 'jbuilder', '~> 2.0'
-# bundle exec rake doc:rails generates the API under doc/api.
-gem 'sdoc', '~> 0.4.0', group: :doc
+gem 'json', '2.3.1'
 
 # Use ActiveModel has_secure_password
 gem 'bcrypt', '~> 3.1.7'
 
 # Used for html parsing
 gem 'nokogiri'
-gem 'nokogumbo'
 gem 'rest-client'
 
 gem 'tzinfo'
 gem 'tzinfo-data'
 gem 'sprockets-rails'
-gem 'pg', '~> 0.20'
+gem 'sprockets', '~> 3.7.2'
+gem 'pg'
 
 # Use Unicorn as the app server
 # gem 'unicorn'
@@ -44,12 +43,7 @@ gem 'pg', '~> 0.20'
 
 group :development, :test do
   # Call 'byebug' anywhere in the code to stop execution and get a debugger console
-  # gem 'byebug'
-end
-
-group :development do
-  # Access an IRB console on exception pages or by using <%= console %> in views
-  gem 'web-console', '~> 2.0'
+  gem 'byebug'
 end
 
 gem 'sprockets_better_errors'
