@@ -1,8 +1,8 @@
 class Match < ActiveRecord::Base
   belongs_to :fixture
 
-  belongs_to :home_team, class_name: Team, :foreign_key => "home_team_id"
-  belongs_to :away_team, class_name: Team, :foreign_key => "away_team_id"
+  belongs_to :home_team, class_name: Team.to_s, :foreign_key => "home_team_id"
+  belongs_to :away_team, class_name: Team.to_s, :foreign_key => "away_team_id"
 
   def home_info
     home_team.name
