@@ -1,4 +1,4 @@
 class Bet < ActiveRecord::Base
-  belongs_to :match
-  belongs_to :user_bet
+  belongs_to :match, :class_name => Match.to_s
+  belongs_to :user_bet, :class_name => UserBet.to_s
 end
